修复issue13，在非管理员的搜索结果中不显示管理列

+++ src/components/SearchResultTable.tsx
@@ -103,28 +103,30 @@ export default function SearchResultTable({
         render: (value: any) => value?.toString() || '-'
     }));
 
-    // 添加管理列
-    const managementColumn: TableColumnsType[number] = {
-        title: getResultTableItemName('manage'),
-        key: 'manage',
-        fixed: 'right',
-        width: 100,
-        render: (_, record) => (
-            (isSystemAdmin || allContestOfficial.includes(record.mid)) && <div style={{ display: 'flex', alignItems: 'center', gap: 1 }}>
-                <ResultEditForm defaultValues={record} isEditMode onSuccess={onContentReFresh} frozenItems={["meet", "projectname", "name", "groupname"]}/>
-                <Button
-                    type="link"
-                    onClick={() => handleDeleteClick({...record} as DeleteRequest)}
-                    style={{ padding: 0 }}
-                    icon={<DeleteOutlined />}
-                >
-                    删除
-                </Button>
-            </div>
-        )
-    };
-
-    const columns = [...baseColumns, managementColumn];
+    let columns: TableColumnsType<any> = [...baseColumns];
+    if (isSystemAdmin || allContestOfficial.length > 0) {
+        // 添加管理列
+        const managementColumn: TableColumnsType[number] = {
+            title: getResultTableItemName('manage'),
+            key: 'manage',
+            fixed: 'right',
+            width: 100,
+            render: (_, record) => (
+                (isSystemAdmin || allContestOfficial.includes(record.mid)) && <div style={{ display: 'flex', alignItems: 'center', gap: 1 }}>
+                    <ResultEditForm defaultValues={record} isEditMode onSuccess={onContentReFresh} frozenItems={["meet", "projectname", "name", "groupname"]}/>
+                    <Button
+                        type="link"
+                        onClick={() => handleDeleteClick({...record} as DeleteRequest)}
+                        style={{ padding: 0 }}
+                        icon={<DeleteOutlined />}
+                    >
+                        删除
+                    </Button>
+                </div>
+            )
+        };
+        columns = [...columns, managementColumn]; 
+    }
 
     // 处理数据源
     const dataSource = results.map((r, index) => ({
